+++ NEWS.md
@@ -1,7 +1,8 @@
 # mpindex 0.2.1
 
 * Changed returned value of `define_mpi_specs` from list to data frame with a class of `mpi_specs_df`. By default, it stores a global option named `mpi_specs` which can be accessed using `getOption('mpi_specs')`. Other specifications are now stored as `attributes`.
+* Changed default value of `.unit_of_analysis` from `households` to `NULL`.
+* `save_mpi` now allows unformatted output (tidy format) when generating an output. 
+* Implemented feature to include MPI specification in the output by setting `.include_specs` tot `TRUE`.
 * Deprecated `.names_separator` argument in `define_mpi_specs`.
-* Allow unformatted output (tidy format) when using `save_mpi`. 
 * Added `use_global_mpi_specs()` wrapper function to use Global MPI specification.
-* Added a `NEWS.md` file to track changes to the package.